server/requirements.txt
alembic==1.4.2
astroid==1.6.6
attrs==19.3.0
backports.functools-lru-cache==1.6.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.8
cssselect==1.1.0
DateTime==4.3
dnspython==1.16.0
enum34==1.1.9
eventlet==0.25.2
flake8==3.8.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-WhooshAlchemy==0.56
futures==3.1.1
geographiclib==1.50
geoip2==3.0.0
geopy==1.21.0
greenlet==0.4.15
gunicorn==19.10.0
http-ece==1.1.0
httpagentparser==1.9.0
idna==2.9
importlib-metadata==1.6.0
ipaddress==1.0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
lxml==4.5.0
Mako==1.1.3
markdown2==2.3.8
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
maxminddb==1.5.4
mccabe==0.6.1
monotonic==1.5
more-itertools==8.4.0
numpy==1.19.0
packaging==20.4
pathlib2==2.3.5
Pillow==6.2.2
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.9.0
py-vapid==1.7.0
pycodestyle==2.6.0
pycparser==2.19
pyflakes==2.2.0
PyJWT==1.7.1
pylint==1.9.5
pyparsing==2.4.7
pyquery==1.4.1
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
pywebpush==1.10.1
readtime==1.1.1
regex==2020.2.20
requests==2.23.0
retinasdk==1.0.0
scandir==1.10.0
singledispatch==3.4.0.3
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
typing==3.7.4.1
urllib3==1.25.8
wcwidth==0.2.5
webptools==0.0.2
Werkzeug==1.0.0
Whoosh==2.7.4
wrapt==1.12.1
zipp==1.2.0
zope.interface==4.7.1
